bigint

+++ db/migrate/20231018054931_create_pug_networks.rb
@@ -1,7 +1,7 @@
 class CreatePugNetworks < ActiveRecord::Migration[7.1]
   def change
     create_table :pug_networks do |t|
-      t.integer :chain_id
+      t.bigint :chain_id
       t.string :name
       t.string :display_name
       t.json :rpc_list
